requirements.txt
flask~=2.1.2
feedparser~=6.0.8
python-dotenv~=0.19.2
beautifulsoup4~=4.10.0
pyyaml~=6.0
requests~=2.27.1
